fixed issues with models, rebuilt DB, added update history, and Role seed data

+++ taskboard-api/Services/IssueService/IssueService.cs
@@ -4,6 +4,7 @@ using System.Security.Claims;
 using taskboard_api.Data;
 using taskboard_api.DTOs.Issue;
 using taskboard_api.DTOs.User;
+using taskboard_api.DTOs.UserRole;
 using taskboard_api.Models;
 
 namespace taskboard_api.Services.IssueService
@@ -23,55 +24,88 @@ namespace taskboard_api.Services.IssueService
             _httpContextAccessor = httpContextAccessor;
         }
 
-        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext!.User
+        private int GetCurrentUserId() => int.Parse(_httpContextAccessor.HttpContext!.User
             .FindFirstValue(ClaimTypes.NameIdentifier)!);
 
+        private async Task<bool> IsValidUserId(int Id)
+        {
+            if (await _context.Users.AnyAsync(u => u.Id == Id))
+            {
+                return true;
+            }
+            return false;
+        }
         public async Task<ServiceResponse<List<GetIssueDTO>>> AddIssue(AddIssueDTO newIssue, int submittedBy)
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
-            var submittingResponse = new ServiceResponse<User>();
-            var assignedResponse = new ServiceResponse<User>();
-            //Issue issue = _mapper.Map<Issue>(newIssue);
-
-            submittingResponse = await _authRepo.GetUser(submittedBy);
-            User userSubmitting = submittingResponse.Data;
-            User userAssignedTo = null;
-
-            if (newIssue.AssignedToId != null)
+            try
             {
-                assignedResponse = await _authRepo.GetUser((int)newIssue.AssignedToId);
-                userAssignedTo = assignedResponse.Data;
+                var issue = _mapper.Map<Issue>(newIssue);
+                var submittedById = GetCurrentUserId();
+
+                if (submittedById > 0)
+                {
+                    issue.SubmittedById = GetCurrentUserId();
+                }
+                else
+                {
+                    throw new Exception("Unable to find UserId for current user");
+                }
+
+
+                if ((newIssue.AssignedToId != null && await IsValidUserId((int)newIssue.AssignedToId)) || newIssue.AssignedToId == null)
+                {
+                    issue.AssignedToId = newIssue.AssignedToId;
+                }
+                else
+                {
+                    throw new Exception("Unable to add new issue. Assigned to invalid user.");
+                }
+
+                issue.LastUpdatedById = GetCurrentUserId();
+                issue.LastUpdatedDate = DateTime.Now;
+
+                _context.Issues.Add(issue);
+                await _context.SaveChangesAsync();
+
+                //returns list of all of the user's issues
+                serviceResponse.Data =
+                    await _context.Issues
+                    .Where(i => i.SubmittedById == GetCurrentUserId())
+                    .Select(i => _mapper.Map<GetIssueDTO>(i))
+                    .ToListAsync();
+
             }
-            
-            Issue issue = new Issue()
+            catch (Exception ex)
             {
-                Title = newIssue.Title,
-                Type = newIssue.Type,
-                Priority = newIssue.Priority,
-                Status = newIssue.Status,
-                Description = newIssue.Description,
-                SubmittedBy = userSubmitting,
-                AssignedTo = userAssignedTo
-            };
+                serviceResponse.Success = false;
+                serviceResponse.Message = ex.Message;
+            }
 
-            _context.Issues.Add(issue);
-            await _context.SaveChangesAsync();
-            serviceResponse.Data = await _context.Issues
-                .Select(i => _mapper.Map<GetIssueDTO>(i)).ToListAsync(); 
-            
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<List<GetIssueDTO>>> DeleteIssue(int issueId)
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
-
+            //TODO: add logic for PM and Admin (delete anything)
             try
             {
-                Issue issue = await _context.Issues.FirstAsync(i => i.IssueId == issueId);
+                var issue = await _context.Issues
+                    .FirstOrDefaultAsync(i => i.IssueId == issueId && i.SubmittedById == GetCurrentUserId());
+
+                if (issue is null)
+                {
+                    throw new Exception($"Issue not found. IssueId: {issueId}.");
+                }
+
                 _context.Issues.Remove(issue);
                 await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Issues.Select(i => _mapper.Map<GetIssueDTO>(i)).ToList();
+
+                serviceResponse.Data =
+                    await _context.Issues
+                    .Where(i => i.SubmittedById == GetCurrentUserId())
+                    .Select(i => _mapper.Map<GetIssueDTO>(i)).ToListAsync();
             }
             catch (Exception ex)
             {
@@ -86,18 +120,18 @@ namespace taskboard_api.Services.IssueService
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
             var dbIssues = await _context.Issues
-                .Where(c => c.SubmittedBy.Id == GetUserId())
+                .Where(c => c.SubmittedById == GetCurrentUserId())
                 .ToListAsync();
             serviceResponse.Data = dbIssues.Select(i => _mapper.Map<GetIssueDTO>(i)).ToList();
             return serviceResponse;
-            
+
         }
 
         public async Task<ServiceResponse<List<GetIssueDTO>>> GetAssignedIssues(int userId)
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
             var dbIssues = await _context.Issues
-                .Where(c => c.AssignedTo.Id == userId)
+                .Where(c => c.AssignedToId == userId)
                 .ToListAsync();
             serviceResponse.Data = dbIssues.Select(i => _mapper.Map<GetIssueDTO>(i)).ToList();
             return serviceResponse;
@@ -107,7 +141,7 @@ namespace taskboard_api.Services.IssueService
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
             var dbIssues = await _context.Issues
-                .Where(c => c.AssignedTo.Id == null)
+                .Where(c => c.AssignedToId == null)
                 .ToListAsync();
             serviceResponse.Data = dbIssues.Select(i => _mapper.Map<GetIssueDTO>(i)).ToList();
             return serviceResponse;
@@ -116,7 +150,10 @@ namespace taskboard_api.Services.IssueService
         public async Task<ServiceResponse<List<GetIssueDTO>>> GetAllIssues()
         {
             var serviceResponse = new ServiceResponse<List<GetIssueDTO>>();
-            var dbIssues = await _context.Issues.ToListAsync();
+            var dbIssues = await _context.Issues
+                    //.Include(i => i.SubmittedById)
+                    //    .Include(i => i.AssignedToId)
+                    .ToListAsync();
             serviceResponse.Data = dbIssues.Select(i => _mapper.Map<GetIssueDTO>(i)).ToList();
             return serviceResponse;
 
@@ -126,6 +163,11 @@ namespace taskboard_api.Services.IssueService
         {
             var serviceResponse = new ServiceResponse<GetIssueDTO>();
             var dbIssue = await _context.Issues.FirstOrDefaultAsync(i => i.IssueId == id);
+
+            //....only retrieve issue if user submitted it
+            //var dbIssue = await _context.Issues
+            //    .FirstOrDefaultAsync(i => i.IssueId == id && i.SubmittedBy.Id == GetCurrentUserId());
+
             serviceResponse.Data = _mapper.Map<GetIssueDTO>(dbIssue);
             return serviceResponse;
         }
@@ -139,7 +181,20 @@ namespace taskboard_api.Services.IssueService
                 var issue = await _context.Issues
                     .FirstOrDefaultAsync(i => i.IssueId == updatedIssue.IssueId);
 
-                _mapper.Map(updatedIssue, issue);
+                if (issue is null)
+                {
+                    throw new Exception($"Issue not found. IssueId: {updatedIssue.IssueId}");
+                }
+
+                issue.Title = updatedIssue.Title;
+                issue.Type = updatedIssue.Type;
+                issue.Priority = updatedIssue.Priority;
+                issue.Status = updatedIssue.Status;
+                issue.Description = updatedIssue.Description;
+                issue.AssignedToId = updatedIssue.AssignedToId;
+                issue.LastUpdatedById = GetCurrentUserId();
+                issue.LastUpdatedDate = DateTime.Now;
+
                 await _context.SaveChangesAsync();
                 serviceResponse.Data = _mapper.Map<GetIssueDTO>(issue);
             }
